Added DBHelper with some tests.

+++ app/src/main/java/es/idenjoe/everpobre/model/Notebook.java
@@ -5,11 +5,11 @@ package es.idenjoe.everpobre.model;
  */
 public class Notebook {
     public static final String DEFAULT_NAME = "John Doe";
-    private int id;
+    private long id;
     private String name;
 
 
-    public Notebook(int id, String name) {
+    public Notebook(long id, String name) {
         if (name == null || name.isEmpty()) {
             this.name = DEFAULT_NAME;
         }else{
@@ -23,11 +23,11 @@ public class Notebook {
         //disable default constructor, customer requirement
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
